Refactor Components and Remove Editor Context
