+++ components/BlueskyPost.tsx
@@ -6,7 +6,7 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
-import { useEditor } from "@/context/EditorContext";
+import { useEditor } from "@/hooks/useEditor";
 import { BlueskyClient } from "@/lib/bluesky";
 import { Switch } from "@/components/ui/switch";
 import {
